Moved Contact DB interface and class to Helpers/contactDB folder
