+++ src/SendGrid/SendGridClient.cs
@@ -16,6 +16,7 @@ using System.Reflection;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
+using SendGrid.Helpers.ContactDb;
 
 namespace SendGrid
 {
